requirements.txt
gate-api>=1.35.0
python-dotenv>=0.19.0
requests>=2.28.0
PyYAML>=6.0

